Move general purpose browser API modules to their own directory
