+++ extension/src/browser/scripting.ts
