+++ extension/src/browser/utils.ts
@@ -1,0 +1,8 @@
+export function buildUrlPatterns(fqdns: string[]): string[] {
+  const urls: string[] = [];
+  for (const fqdn of fqdns) {
+    urls.push(`http://${fqdn}/*`);
+    urls.push(`https://${fqdn}/*`);
+  }
+  return urls;
+}
