+++ extension/src/webcat/utils.ts
@@ -160,12 +160,3 @@ export async function getFirstParty(
   );
   return details.requestId;
 }
-
-export function buildUrlPatterns(fqdns: string[]): string[] {
-  const urls: string[] = [];
-  for (const fqdn of fqdns) {
-    urls.push(`http://${fqdn}/*`);
-    urls.push(`https://${fqdn}/*`);
-  }
-  return urls;
-}
